requirements.txt
llama-index-core==0.11.23
pymilvus==2.4.5
